Feature/tests

See merge request miszczu1902/baker-tech!15

+++ backend/src/main/java/pl/lodz/p/it/bakertech/accounts/services/impl/AccountScheduleServiceImpl.java
@@ -15,10 +15,13 @@ import pl.lodz.p.it.bakertech.accounts.services.AccountScheduleService;
 import pl.lodz.p.it.bakertech.common.CommonService;
 import pl.lodz.p.it.bakertech.exceptions.AppException;
 import pl.lodz.p.it.bakertech.interceptors.schedule.ScheduledInterception;
+import pl.lodz.p.it.bakertech.model.accounts.Account;
 import pl.lodz.p.it.bakertech.model.accounts.AccountConfirmationToken;
+import pl.lodz.p.it.bakertech.utils.DateUtility;
 import pl.lodz.p.it.bakertech.utils.mappers.accounts.KeycloakMapper;
 import pl.lodz.p.it.bakertech.validation.etag.ETagGenerator;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 import static pl.lodz.p.it.bakertech.utils.SchedulePeriods.PER_ONE_MINUTE;
@@ -53,17 +56,22 @@ public class AccountScheduleServiceImpl extends CommonService implements Account
         List<AccountConfirmationToken> unconfirmedAccounts = accountConfirmationTokenRepository.findAll();
         if (!unconfirmedAccounts.isEmpty()) {
             unconfirmedAccounts.forEach(accountConfirmationToken -> {
-                String username = accountConfirmationToken.getAccount().getUsername();
+                Account account = accountConfirmationToken.getAccount();
+                String username = account.getUsername();
+                LocalDateTime now = DateUtility.nowWithTimestamp();
+                LocalDateTime removalDate = account.getCreationDateTime().plusDays(1);
 
-                accountRepository.delete(accountConfirmationToken.getAccount());
-                accountConfirmationTokenRepository.delete(accountConfirmationToken);
+                if (now.isAfter(removalDate) || now.isEqual(removalDate)) {
+                    accountRepository.delete(account);
+                    accountConfirmationTokenRepository.delete(accountConfirmationToken);
 
-                try (Response response = realmResource.users().delete(getKeycloakUserByUsername(username).getId())) {
-                    if (response.getStatus() != 204 && response.getStatus() != 200) {
-                        throw AppException.createKeycloakException();
+                    try (Response response = realmResource.users().delete(getKeycloakUserByUsername(username).getId())) {
+                        if (response.getStatus() != 204 && response.getStatus() != 200) {
+                            throw AppException.createKeycloakException();
+                        }
+                    } catch (Exception e) {
+                        throw AppException.createKeycloakException(e.getCause());
                     }
-                } catch (Exception e) {
-                    throw AppException.createKeycloakException(e.getCause());
                 }
             });
         }
